Refresh-Spotify-Access as a function that you can call it before every Spotify API call
Refresh Token on Path Change

Remove param after spotify login page
